# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Company-in-a-Box API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  draft: 'draft',
  running: 'running',
  completed: 'completed',
} as const;

export interface Project {
  id: number;
  name: string;
  businessIdea: string;
  /** @nullable */
  industry: string | null;
  /** @nullable */
  targetMarket: string | null;
  status: ProjectStatus;
  completionPercent: number;
  createdAt: string;
}

export interface ProjectInput {
  name: string;
  businessIdea: string;
  industry?: string;
  targetMarket?: string;
}

export type ProjectWithTasksStatus = typeof ProjectWithTasksStatus[keyof typeof ProjectWithTasksStatus];


export const ProjectWithTasksStatus = {
  draft: 'draft',
  running: 'running',
  completed: 'completed',
} as const;

export type AgentTaskAgentType = typeof AgentTaskAgentType[keyof typeof AgentTaskAgentType];


export const AgentTaskAgentType = {
  ceo: 'ceo',
  marketing: 'marketing',
  sales: 'sales',
  cskh: 'cskh',
  hr: 'hr',
  accountant: 'accountant',
  legal: 'legal',
} as const;

export type AgentTaskStatus = typeof AgentTaskStatus[keyof typeof AgentTaskStatus];


export const AgentTaskStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface AgentTask {
  id: number;
  projectId: number;
  agentType: AgentTaskAgentType;
  agentName: string;
  status: AgentTaskStatus;
  /** @nullable */
  output: string | null;
  /** @nullable */
  errorMessage: string | null;
  createdAt: string;
  /** @nullable */
  completedAt: string | null;
}

export interface ProjectWithTasks {
  id: number;
  name: string;
  businessIdea: string;
  /** @nullable */
  industry: string | null;
  /** @nullable */
  targetMarket: string | null;
  status: ProjectWithTasksStatus;
  completionPercent: number;
  createdAt: string;
  tasks: AgentTask[];
}

export interface ProjectSummary {
  projectId: number;
  projectName: string;
  businessIdea: string;
  totalAgents: number;
  completedAgents: number;
  runningAgents: number;
  pendingAgents: number;
  failedAgents: number;
  completionPercent: number;
  status: string;
  tasks: AgentTask[];
}

export interface RunAllResult {
  projectId: number;
  agentsTriggered: string[];
  message: string;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface ApiError {
  error: string;
}