# lib/api-zod/src/generated/types/agentTask.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Company-in-a-Box API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentTaskAgentType } from './agentTaskAgentType';
import type { AgentTaskStatus } from './agentTaskStatus';

export interface AgentTask {
  id: number;
  projectId: number;
  agentType: AgentTaskAgentType;
  agentName: string;
  status: AgentTaskStatus;
  /** @nullable */
  output: string | null;
  /** @nullable */
  errorMessage: string | null;
  createdAt: Date;
  /** @nullable */
  completedAt: string | null;
}