# lib/api-zod/src/generated/types/projectInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Company-in-a-Box API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectInput {
  name: string;
  businessIdea: string;
  industry?: string;
  targetMarket?: string;
}