# lib/api-zod/src/generated/types/projectSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Company-in-a-Box API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentTask } from './agentTask';

export interface ProjectSummary {
  projectId: number;
  projectName: string;
  businessIdea: string;
  totalAgents: number;
  completedAgents: number;
  runningAgents: number;
  pendingAgents: number;
  failedAgents: number;
  completionPercent: number;
  status: string;
  tasks: AgentTask[];
}