# lib/api-zod/src/generated/types/projectWithTasks.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Company-in-a-Box API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentTask } from './agentTask';
import type { ProjectWithTasksStatus } from './projectWithTasksStatus';

export interface ProjectWithTasks {
  id: number;
  name: string;
  businessIdea: string;
  /** @nullable */
  industry: string | null;
  /** @nullable */
  targetMarket: string | null;
  status: ProjectWithTasksStatus;
  completionPercent: number;
  createdAt: Date;
  tasks: AgentTask[];
}